Tests/LibWeb/Text/expected/DOM/Element-classList.txt

 element.classList initial value: ""
element.classList after setting classList to "a": "a"
element.classList after setting className to "": ""
